+++ src/modules/notification/notification.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { LoggerModule } from '../logger/logger.module';
+import { NotificationService } from './notification.service';
+
+@Module({
+    imports: [LoggerModule],
+    providers: [NotificationService],
+    exports: [NotificationService],
+})
+export class NotificationModule {}